+++ src/com/jpexs/games/rpsls/StartLocalGameFrame.java
@@ -3,7 +3,6 @@ package com.jpexs.games.rpsls;
 import com.jpexs.games.rpsls.model.GameType;
 import java.awt.Container;
 import java.awt.FlowLayout;
-import java.awt.HeadlessException;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
@@ -21,10 +20,10 @@ import javax.swing.JTextField;
  *
  * @author JPEXS
  */
-public class StartNetworkGameFrame extends JFrame {
+public class StartLocalGameFrame extends JFrame {
 
-    public StartNetworkGameFrame() {
-        setTitle("RPSLS - Start network game");
+    public StartLocalGameFrame() {
+        setTitle("RPSLS - Start local game");
         addWindowListener(new WindowAdapter() {
             @Override
             public void windowClosing(WindowEvent e) {
@@ -33,13 +32,6 @@ public class StartNetworkGameFrame extends JFrame {
         });
         Container container = getContentPane();
         container.setLayout(new BoxLayout(container, BoxLayout.Y_AXIS));
-        JPanel portPanel = new JPanel(new FlowLayout());
-        JTextField portField = new JTextField("1024", 5);
-
-        portPanel.add(new JLabel("Port:"));
-        portPanel.add(portField);
-        container.add(portPanel);
-
         JLabel gameTypeLabel = new JLabel("Game type:");
         gameTypeLabel.setAlignmentX(0.5f);
         container.add(gameTypeLabel);
@@ -51,14 +43,8 @@ public class StartNetworkGameFrame extends JFrame {
         okButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                try {
-                    int port = Integer.parseInt(portField.getText());
-                    setVisible(false);
-                    Main.startNetworkGame(port, (GameType) gameTypeComboBox.getSelectedItem());
-
-                } catch (NumberFormatException exception) {
-                    JOptionPane.showMessageDialog(StartNetworkGameFrame.this, "Invalid port", "Error", JOptionPane.ERROR_MESSAGE);
-                }
+                setVisible(false);
+                Main.startLocalGame((GameType) gameTypeComboBox.getSelectedItem());
             }
         });
 
@@ -80,5 +66,4 @@ public class StartNetworkGameFrame extends JFrame {
         Main.centerWindow(this);
         Main.setWindowIcon(this);
     }
-
 }